+++ packages/@expo/metro-config/src/ExpoMetroConfig.ts
@@ -84,9 +84,6 @@ export function getDefaultConfig(
   const sourceExtsConfig = { isTS: true, isReact: true, isModern: true };
   const sourceExts = getBareExtensions([], sourceExtsConfig);
 
-  // Add support for cjs (without platform extensions).
-  sourceExts.push('cjs');
-
   const reanimatedVersion = getPkgVersion(projectRoot, 'react-native-reanimated');
 
   let sassVersion: string | null = null;
@@ -146,7 +143,7 @@ export function getDefaultConfig(
     },
     watcher: {
       // strip starting dot from env files
-      additionalExts: envFiles.map((file: string) => file.replace(/^\./, '')),
+      additionalExts: ['cjs', 'mjs', ...envFiles.map((file: string) => file.replace(/^\./, ''))],
     },
     serializer: {
       getModulesRunBeforeMainModule: () => {